code/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-jet-reboot==1.3.3
django-ninja==0.22.2
feedparser==6.0.10
pydantic==1.10.12
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
